add package publish tag:
