+++ package.json
@@ -3,7 +3,7 @@
   "version": "0.5.1",
   "private": false,
   "description": "Dipperin javascript lib",
-  "main": "build/web/dipperin.js",
+  "main": "build/main/index.js",
   "typings": "build/module/index.d.ts",
   "module": "build/module/index.js",
   "license": "MIT",
@@ -12,8 +12,10 @@
     "info": "npm-scripts-info",
     "dev": "yarn clean && tsc -p tsconfig.watch.json",
     "build": "yarn clean && yarn build:module && yarn build:umd",
+    "build:web": "yarn clean && yarn build:module && yarn build:umdweb",
     "build:module": "tsc -p tsconfig.json",
-    "build:umd": "webpack-cli --config webpack.config.js",
+    "build:umd": "webpack-cli --config webpack.node.config.js",
+    "build:umdweb": "webpack-cli --config webpack.web.config.js",
     "analyze": "source-map-explorer build/main/*.*",
     "prettier": "prettier --write",
     "fix": "yarn fix:prettier && yarn fix:tslint",
